### 2021-06-15 18f7845
Rework contributing docs into new tutorial, renamed 'Use Cases' to 'Getting started'
