### 2021-06-15 c7e7edd
Add some more tutorial-style documentation (#850)
* Add brief descriptions of OPTIMADE use cases

* Add usage of the validator

* Rework contributing docs into new tutorial on filter transformers

* Renamed 'Use Cases' to 'Getting started'
