+++ optimade/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.15.3"
+__version__ = "0.15.4"
 __api_version__ = "1.0.1"
