### 2021-06-15 c848a25
small corrections to PR #850
Processed my suggestions for the text added by 'PR #850 Add some more tutorial-style documentation'.
