### 2021-06-15 cd1df34
processed my suggestions for PR #850
