### 2021-06-16 252bd19
Apply suggestions from code review
added suggestions CasperW which add references to API documentation.

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
