### 2021-06-16 c78dd4c
Add info about the levels of verbosity to --help output.
