### 2021-06-16 8fa3f0c
put text verbose on 1 line
