### 2021-06-16 9632c07
added extra test to tets cases with AND and OR statements with three clauses
