### 2021-06-17 fbe469c
Apply suggestions from code review
Changed formatting description handle_not_and function

Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
