+++ optimade/filtertransformers/mongo.py
@@ -333,9 +333,7 @@ class MongoTransformer(BaseTransformer):
             return subdict
 
         return recursive_postprocessing(
-            filter_,
-            check_for_length_op_filter,
-            apply_length_op,
+            filter_, check_for_length_op_filter, apply_length_op
         )
 
     def _apply_relationship_filtering(self, filter_: dict) -> dict:
@@ -365,11 +363,7 @@ class MongoTransformer(BaseTransformer):
                     subdict["$and"] = []
                 subdict["$and"].extend(
                     [
-                        {
-                            f"relationships.{_prop}.data": {
-                                "$size": expr.pop("$size"),
-                            }
-                        },
+                        {f"relationships.{_prop}.data": {"$size": expr.pop("$size")}},
                         {f"relationships.{_prop}.data.{_field}": expr},
                     ]
                 )
