### 2021-06-17 33353f2
added test to test_filter.py to test not propegation.
