### 2021-06-17 d38cecc
Apply suggestions from code review
Accepted minor changes from code suggestions

Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
