### 2021-06-17 ef5a273
Renamed opposite_operator_map to inverse_operator_map and turned it into a class attribute.
