### 2021-06-18 a4d3424
Documentation tweaks (#852)
Small corrections to PR #850 .

Add info about the levels of verbosity to --help output.


Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
