### 2021-06-18 c3efa26
Remove old stuff + one more test
