+++ tests/filtertransformers/test_ase.py
@@ -16,7 +16,9 @@ def transform():
 
 cases = [
     ('nelements=2',
-     ('nelements', ('=', 2)))]
+     ('nelements', ('=', 2))),
+    ('nelements=3 AND elements HAS "Cu"',
+     ('AND', [('nelements', ('=', 3)), ('elements', ('HAS', '"Cu"'))]))]
 
 
 @pytest.mark.parametrize(['query', 'tree'], cases)
