+++ optimade/server/entry_collections/ase.py
@@ -12,6 +12,7 @@ from optimade.filtertransformers.ase import ASEQueryTree, ASETransformer
 from optimade.models import EntryResource
 from optimade.server.config import CONFIG
 from optimade.server.entry_collections import EntryCollection
+
 # from optimade.server.logger import LOGGER
 from optimade.server.mappers import BaseResourceMapper
 
@@ -19,15 +20,15 @@ INF = 9999999999999
 
 
 class ASECollection(EntryCollection):
-    def __init__(self,
-                 name: str,
-                 resource_cls: EntryResource,
-                 resource_mapper: BaseResourceMapper,
-                 database: str = CONFIG.ase_database):
+    def __init__(
+        self,
+        name: str,
+        resource_cls: EntryResource,
+        resource_mapper: BaseResourceMapper,
+        database: str = CONFIG.ase_database,
+    ):
 
-        super().__init__(resource_cls,
-                         resource_mapper,
-                         ASETransformer())
+        super().__init__(resource_cls, resource_mapper, ASETransformer())
 
         self.parser = LarkParser(version=(1, 0, 0), variant="default")
         self.collection = ASEDBWrapper(database)
@@ -43,19 +44,20 @@ class ASECollection(EntryCollection):
         filter_function = create_filter_function(filter)
         return self.collection.count(filter_function, skip=skip, limit=limit)
 
-    def _run_db_query(self,
-                      criteria: Dict[str, Any],
-                      single_entry: bool = False
-                      ) -> Tuple[List[Dict[str, Any]], int, bool]:
+    def _run_db_query(
+        self, criteria: Dict[str, Any], single_entry: bool = False
+    ) -> Tuple[List[Dict[str, Any]], int, bool]:
 
-        filter_function = create_filter_function(criteria['filter'])
+        filter_function = create_filter_function(criteria["filter"])
 
-        ids = [row.id
-               for row in self.collection.select(filter_function,
-                                                 skip=criteria.get('skip', 0))]
+        ids = [
+            row.id
+            for row in self.collection.select(
+                filter_function, skip=criteria.get("skip", 0)
+            )
+        ]
 
-        results = [self.collection.get_result(id)
-                   for id in ids[:criteria['limit']]]
+        results = [self.collection.get_result(id) for id in ids[: criteria["limit"]]]
 
         if single_entry:
             data_returned = len(results)
@@ -83,11 +85,9 @@ FilterFunction = Callable[[RowInfo], bool]
 def dbrow2rowinfo(row: AtomsRow) -> RowInfo:
     """Pull out stuff needed for search."""
     formula = Formula(row.formula)
-    rowinfo = RowInfo(row.id,
-                      row.count_atoms(),
-                      f'{formula:hill}',
-                      f'{formula:abc}',
-                      row.natoms)
+    rowinfo = RowInfo(
+        row.id, row.count_atoms(), f"{formula:hill}", f"{formula:abc}", row.natoms
+    )
 
     return rowinfo
 
@@ -101,10 +101,9 @@ class ASEDBWrapper:
     def __len__(self):
         return len(self.rows)
 
-    def select(self,
-               filter_function: FilterFunction,
-               skip: int = 0,
-               limit: int = INF) -> Generator[RowInfo, None, None]:
+    def select(
+        self, filter_function: FilterFunction, skip: int = 0, limit: int = INF
+    ) -> Generator[RowInfo, None, None]:
         """Select rows using a filter function."""
         for i, row in enumerate(self.rows):
             if i == skip + limit:
@@ -113,9 +112,7 @@ class ASEDBWrapper:
                 yield row
 
     def count(self, filter_function, skip=0, limit=INF):
[…]
