### 2021-06-18 846a7dc
added test for double negation : {'': {''
