### 2021-06-18 8ebe7ad
added test for double negation : {'': {''
