### 2021-06-21 1e5e79a
Updated INSTALL.md with with instructions to retrieve the list of providers.
