### 2021-06-22 f13a337
reworded instructions for retrieving optimade providers
Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
