+++ INSTALL.md
@@ -38,7 +38,7 @@ conda activate optimade
 # Install package and dependencies in editable mode (including "dev" requirements).
 pip install -e ".[dev]"
 
-# Retrieve a list of providers. (Without this list some of the test will fail because providers.json cannot be found.)
+# Optional: Retrieve the list of OPTIMADE providers. (Without this submodule, some of the tests will fail because "providers.json" cannot be found.)
 git submodule update --init
 
 # Run the tests with pytest
