### 2021-06-23 2957b2b
Update to AiiDA v1.6.4 and remove CI fix
