### 2021-06-23 be8182b
added extra tests to test behaviour NOT operator.
