### 2021-06-24 5690e1e
Update to AiiDA v1.6.4 and remove CI fix (#857)
Fix test utility function for version retrieval.
Add support for commas (`,`) in the version.
Always choose the "first" version given.
