+++ requirements-client.txt
@@ -1,4 +1,4 @@
-aiida-core==1.6.3
+aiida-core==1.6.4
 ase==3.21.1
 numpy==1.20.3
 pymatgen==2021.3.9
