### 2021-06-24 16fc3e0
Added code to correctly handle HAS NOT queries on references. It is however rather ugly as I had to hard code it.
