### 2021-06-24 b3a6a73
storing work in progress on queries on speciesfields
