### 2021-06-25 e1623c6
Updated comments in code with suggestions Matthew
Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
