### 2021-06-25 c2a1a37
removed obsolete code for handling HAS ONLY
