+++ optimade/filtertransformers/mongo.py
@@ -362,19 +362,7 @@ class MongoTransformer(BaseTransformer):
                     f'Cannot filter relationships by field "{_field}", only "id" is supported.'
                 )
 
-            # in the case of HAS ONLY, the size operator needs to be applied
-            # one level up, i.e. excluding the field
-            if "$size" in expr:
-                if "$and" not in subdict:
-                    subdict["$and"] = []
-                subdict["$and"].extend(
-                    [
-                        {f"relationships.{_prop}.data": {"$size": expr.pop("$size")}},
-                        {f"relationships.{_prop}.data.{_field}": expr},
-                    ]
-                )
-            else:
-                subdict[f"relationships.{_prop}.data.{_field}"] = expr
+            subdict[f"relationships.{_prop}.data.{_field}"] = expr
 
             subdict.pop(prop)
 
