### 2021-06-25 8685e40
Applied coode suggestions for comments from Matthew
Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
