### 2021-06-25 9d06214
Double negations, e.g. NOT(NOT(... , is now simplified correctly.
