+++ tests/filtertransformers/test_mongo.py
@@ -148,18 +148,20 @@ class TestMongoTransformer:
         }
 
         assert self.transform('NOT(NOT(chemical_formula_hill CONTAINS "Al"))') == {
-            "$and": [
-                {"chemical_formula_hill": {"$regex": "Al"}},
-                {"chemical_formula_hill": {"$ne": None}},
-            ]
+            "chemical_formula_hill": {"$regex": "Al"}
         }
 
         assert self.transform(
             'NOT(nelements > 3 AND NOT(elements HAS "Ac" AND chemical_formula_hill CONTAINS "Al"))'
         ) == {
-            "$and": [
-                {"chemical_formula_hill": {"$not": {"$not": {"$regex": "Al"}}}},
-                {"chemical_formula_hill": {"$ne": None}},
+            "$or": [
+                {"nelements": {"$lte": 3}},
+                {
+                    "$and": [
+                        {"elements": {"$in": ["Ac"]}},
+                        {"chemical_formula_hill": {"$regex": "Al"}},
+                    ]
+                },
             ]
         }
 
