### 2021-06-29 759412a
Fix query negation in MongoDB (#814)
* Test now expects NOT statements to be simplified and propagated and also requires checks for the existence of properties in case of  and  .

* fixed typo in test_operators in test_mongo.py

* Fix issue_#79, added not propagation and simplifies queries with not by using the opposite operator. The issue with not selecting entries where the property is missing or has value null however remains.

* Documents with missing properties or a value of null for the queried property are no longer returned.

* Double negations now remove "prop" : {"$ne" : None} field.
[…]
