+++ tests/filtertransformers/test_mongo.py
@@ -107,27 +107,96 @@ class TestMongoTransformer:
 
     def test_operators(self):
         # Basic boolean operations
-        # TODO: {"a": {"$not": {"$lt": 3}}} can be simplified to {"a": {"$gte": 3}}
-        assert self.transform("NOT a<3") == {"a": {"$not": {"$lt": 3}}}
+        assert self.transform("NOT a<3") == {"a": {"$gte": 3}}
 
-        # TODO: {'$not': {'$eq': 'Ti'}} can be simplified to {'$ne': 'Ti'}
+        assert self.transform('NOT ( chemical_formula_hill = "Al")') == {
+            "$and": [
+                {"chemical_formula_hill": {"$ne": "Al"}},
+                {"chemical_formula_hill": {"$ne": None}},
+            ]
+        }
+
+        assert self.transform("NOT( NOT( a=4 OR b= 6))") == {
+            "$or": [{"a": {"$eq": 4}}, {"b": {"$eq": 6}}]
+        }
+
+        assert self.transform("a=4 AND b= 6 AND c=10") == {
+            "$and": [{"a": {"$eq": 4}}, {"b": {"$eq": 6}}, {"c": {"$eq": 10}}]
+        }
+
+        assert self.transform("NOT(a<4 OR a>20 AND NOT(c=10))") == {
+            "$and": [
+                {"a": {"$gte": 4}},
+                {"$or": [{"a": {"$lte": 20}}, {"c": {"$eq": 10}}]},
+            ]
+        }
+
+        assert self.transform("NOT(a>4 AND a<20 AND c=10)") == {
+            "$or": [
+                {"a": {"$lte": 4}},
+                {"a": {"$gte": 20}},
+                {"$and": [{"c": {"$ne": 10}}, {"c": {"$ne": None}}]},
+            ]
+        }
+
+        assert self.transform("NOT(a>4 OR a<20 OR c=10)") == {
+            "$and": [
+                {"a": {"$lte": 4}},
+                {"a": {"$gte": 20}},
+                {"$and": [{"c": {"$ne": 10}}, {"c": {"$ne": None}}]},
+            ]
+        }
+
+        assert self.transform('NOT(NOT(chemical_formula_hill CONTAINS "Al"))') == {
+            "chemical_formula_hill": {"$regex": "Al"}
+        }
+
+        assert self.transform(
+            'NOT(nelements > 3 AND NOT(elements HAS "Ac" AND chemical_formula_hill CONTAINS "Al"))'
+        ) == {
+            "$or": [
+                {"nelements": {"$lte": 3}},
+                {
+                    "$and": [
+                        {"elements": {"$in": ["Ac"]}},
+                        {"chemical_formula_hill": {"$regex": "Al"}},
+                    ]
+                },
+            ]
+        }
+
         assert self.transform(
             "NOT ( "
             'chemical_formula_hill = "Al" AND chemical_formula_anonymous = "A" OR '
             'chemical_formula_anonymous = "H2O" AND NOT chemical_formula_hill = "Ti" '
             ")"
         ) == {
-            "$nor": [
+            "$and": [
                 {
-                    "$and": [
-                        {"chemical_formula_hill": {"$eq": "Al"}},
-                        {"chemical_formula_anonymous": {"$eq": "A"}},
+                    "$or": [
+                        {
+                            "$and": [
+                                {"chemical_formula_hill": {"$ne": "Al"}},
+                                {"chemical_formula_hill": {"$ne": None}},
+                            ]
+                        },
+                        {
+                            "$and": [
+                                {"chemical_formula_anonymous": {"$ne": "A"}},
+                                {"chemical_formula_anonymous": {"$ne": None}},
+                            ]
+                        },
                     ]
                 },
                 {
-                    "$and": [
[…]
