### 2021-06-30 5d3d30e
Corrected the behaviour of HAS ONLY and updated tests with the correct results.
