+++ tests/server/query_params/test_filter.py
@@ -185,7 +185,7 @@ def test_list_length_comparisons_unaliased(check_response, check_error_response)
 def test_list_has_only(check_response):
     """Test HAS ONLY query on elements."""
     request = '/structures?filter=elements HAS ONLY "Ac", "Mg"'
-    expected_ids = ["mpf_23"]
+    expected_ids = ["mpf_1", "mpf_23"]
     check_response(request, expected_ids)
 
     request = '/structures?filter=elements HAS ONLY "Ac"'
