+++ tests/filtertransformers/test_mongo.py
@@ -376,8 +376,12 @@ class TestMongoTransformer:
 
         assert t.transform(p.parse('structures.id HAS ONLY "dummy/2019"')) == {
             "$and": [
-                {"relationships.structures.data": {"$size": 1}},
-                {"relationships.structures.data.id": {"$all": ["dummy/2019"]}},
+                {
+                    "relationships.structures.data.id": {
+                        "$not": {"$elemMatch": {"$nin": ["dummy/2019"]}}
+                    }
+                },
+                {"relationships.structures.data.id.0": {"$exists": True}},
             ]
         }
 
@@ -390,12 +394,12 @@ class TestMongoTransformer:
                 {
                     "$and": [
                         {
-                            "relationships.structures.data": {
-                                "$size": 1,
+                            "relationships.structures.data.id": {
+                                "$not": {"$elemMatch": {"$nin": ["dummy/2019"]}}
                             }
                         },
-                        {"relationships.structures.data.id": {"$all": ["dummy/2019"]}},
-                    ],
+                        {"relationships.structures.data.id.0": {"$exists": True}},
+                    ]
                 },
                 {"relationships.structures.data.id": {"$in": ["dummy/2019"]}},
             ]
@@ -694,7 +698,14 @@ class TestMongoTransformer:
     def test_list_properties(self):
         """Test the HAS ALL, ANY and optional ONLY queries."""
         assert self.transform('elements HAS ONLY "H","He","Ga","Ta"') == {
-            "elements": {"$all": ["H", "He", "Ga", "Ta"], "$size": 4}
+            "$and": [
+                {
+                    "elements": {
+                        "$not": {"$elemMatch": {"$nin": ["H", "He", "Ga", "Ta"]}}
+                    }
+                },
+                {"elements.0": {"$exists": True}},
+            ]
         }
 
         assert self.transform('elements HAS ANY "H","He","Ga","Ta"') == {
@@ -713,10 +724,16 @@ class TestMongoTransformer:
                 {"elements": {"$in": ["H"]}},
                 {"elements": {"$all": ["H", "He", "Ga", "Ta"]}},
                 {
-                    "elements": {
-                        "$all": ["H", "He", "Ga", "Ta"],
-                        "$size": 4,
-                    }
+                    "$and": [
+                        {
+                            "elements": {
+                                "$not": {
+                                    "$elemMatch": {"$nin": ["H", "He", "Ga", "Ta"]}
+                                }
+                            }
+                        },
+                        {"elements.0": {"$exists": True}},
+                    ]
                 },
                 {"elements": {"$in": ["H", "He", "Ga", "Ta"]}},
             ]
