+++ optimade/validator/utils.py
@@ -162,13 +162,17 @@ class ValidatorResults:
 
 class Client:  # pragma: no cover
     def __init__(
-        self, base_url: str, max_retries: int = 5, headers: Dict[str, str] = None
+        self,
+        base_url: str,
+        max_retries: int = 5,
+        headers: Dict[str, str] = None,
+        timeout: Optional[float] = 1.0,
     ) -> None:
         """Initialises the Client with the given `base_url` without testing
         if it is valid.
 
         Parameters:
-            base_url (str): the base URL of the optimade implementation, including
+            base_url: the base URL of the optimade implementation, including
                 request protocol (e.g. `'http://'`) and API version number if necessary.
 
                 Examples:
@@ -180,6 +184,7 @@ class Client:  # pragma: no cover
 
             max_retries: The maximum number of attempts to make for each query.
             headers: Dictionary of additional headers to add to every request.
+            timeout: Request timeout in seconds.
 
         """
         self.base_url = base_url
@@ -187,6 +192,7 @@ class Client:  # pragma: no cover
         self.response = None
         self.max_retries = max_retries
         self.headers = headers or {}
+        self.timeout = timeout or 1.0
 
     def get(self, request: str):
         """Makes the given request, with a number of retries if being rate limited. The
@@ -217,8 +223,11 @@ class Client:  # pragma: no cover
         # probably a smarter way to do this with requests, but their documentation 404's...
         while retries < self.max_retries:
             retries += 1
+            exc = None
             try:
-                self.response = requests.get(self.last_request, headers=self.headers)
+                self.response = requests.get(
+                    self.last_request, headers=self.headers, timeout=self.timeout
+                )
             except requests.exceptions.ConnectionError as exc:
                 sys.exit(
                     f"{exc.__class__.__name__}: No response from server at {self.last_request}, please check the URL."
@@ -227,6 +236,9 @@ class Client:  # pragma: no cover
                 sys.exit(
                     f"Unable to make request on {self.last_request}, did you mean http://{self.last_request}?"
                 )
+            except requests.exceptions.Timeout as e:
+                exc = e
+
             status_code = self.response.status_code
             if status_code != 429:
                 break
@@ -234,7 +246,10 @@ class Client:  # pragma: no cover
             time.sleep(1)
 
         else:
-            raise ResponseError("Hit max (manual) retries on request.")
+            message = "Hit max (manual) retries on request."
+            if exc is not None:
+                message += f" Last error: {exc}"
+            raise ResponseError(message)
 
         return self.response
 
