+++ requirements.txt
@@ -3,7 +3,7 @@ lark-parser==0.11.3
 pydantic==1.8.2
 email_validator==1.1.3
 requests==2.25.1
-uvicorn==0.13.4
+uvicorn==0.14.0
 pymongo==3.11.4
 mongomock==3.23.0
 elasticsearch-dsl==6.4.0
