+++ requirements-docs.txt
@@ -1,4 +1,4 @@
-mkdocs==1.1.2
+mkdocs==1.2.1
 mkdocs-awesome-pages-plugin==2.5.0
 mkdocs-material==7.1.6
 mkdocs-minify-plugin==0.4.0
