+++ requirements-client.txt
@@ -2,4 +2,4 @@ aiida-core==1.6.4
 ase==3.21.1
 numpy==1.20.3
 pymatgen==2021.3.9
-jarvis-tools==2021.5.31
+jarvis-tools==2021.6.18
