### 2021-07-01 ff8d584
Bump ase from 3.21.1 to 3.22.0
Bumps [ase](https://gitlab.com/ase/ase) from 3.21.1 to 3.22.0.
- [Release notes](https://gitlab.com/ase/ase/tags)
- [Changelog](https://gitlab.com/ase/ase/blob/master/CHANGELOG.rst)
- [Commits](https://gitlab.com/ase/ase/compare/3.21.1...3.22.0)

---
updated-dependencies:
- dependency-name: ase
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
