### 2021-07-01 e863af4
Bump mkdocstrings from 0.15.1 to 0.15.2
Bumps [mkdocstrings](https://github.com/mkdocstrings/mkdocstrings) from 0.15.1 to 0.15.2.
- [Release notes](https://github.com/mkdocstrings/mkdocstrings/releases)
- [Changelog](https://github.com/mkdocstrings/mkdocstrings/blob/master/CHANGELOG.md)
- [Commits](https://github.com/mkdocstrings/mkdocstrings/compare/0.15.1...0.15.2)

---
updated-dependencies:
- dependency-name: mkdocstrings
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
