+++ requirements-docs.txt
@@ -2,4 +2,4 @@ mkdocs==1.1.2
 mkdocs-awesome-pages-plugin==2.5.0
 mkdocs-material==7.1.6
 mkdocs-minify-plugin==0.4.0
-mkdocstrings==0.15.1
+mkdocstrings==0.15.2
