### 2021-07-01 c786fc3
Bump email-validator from 1.1.2 to 1.1.3
Bumps [email-validator](https://github.com/JoshData/python-email-validator) from 1.1.2 to 1.1.3.
- [Release notes](https://github.com/JoshData/python-email-validator/releases)
- [Commits](https://github.com/JoshData/python-email-validator/compare/v1.1.2...v1.1.3)

---
updated-dependencies:
- dependency-name: email-validator
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
