+++ requirements.txt
@@ -1,7 +1,7 @@
 fastapi==0.65.2
 lark-parser==0.11.3
 pydantic==1.8.2
-email_validator==1.1.2
+email_validator==1.1.3
 requests==2.25.1
 uvicorn==0.13.4
 pymongo==3.11.4
