### 2021-07-01 e72c698
Apply suggestions from code review
Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
Co-authored-by: Johan Bergsma <JPBergsma@users.noreply.github.com>
