### 2021-07-01 9f962ba
Update setup.py
Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
