### 2021-07-01 b5eda85
Rename _apply_only_filter to _apply_has_only_filter
Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
