### 2021-07-01 6b7c2ec
Update dependencies (#872)
Update dependencies:

* Bump mkdocs from 1.1.2 to 1.2.1
* Bump pylint from 2.8.3 to 2.9.1
* Bump email-validator from 1.1.2 to 1.1.3
* Bump mkdocs-material from 7.1.6 to 7.1.9
* Bump mkdocstrings from 0.15.1 to 0.15.2
* Bump numpy from 1.20.3 to 1.21.0
* Bump jarvis-tools from 2021.5.31 to 2021.6.18
* Bump pytest-cov from 2.12.0 to 2.12.1
* Bump uvicorn from 0.13.4 to 0.14.0
* Bump ase from 3.21.1 to 3.22.0

Updated setup.py accordingly.

Bump pre-commit hook versions.


Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
