### 2021-07-01 b76a2b3
Gave "structures.id" the same behaviour as "references.id" for HAS_ONLY filters."
