### 2021-07-01 8557502
Increase default connection timeout to 3+epsilon for TCP retransmission
