### 2021-07-03 c601ef7
Add request --timeout parameter to validator (#860)
* Add request --timeout parameter to validator

* Add separate (larger) read timeout

* Apply suggestions from code review

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
Co-authored-by: Johan Bergsma <JPBergsma@users.noreply.github.com>

* Fix old typo in error message that was causing internal errors

* Add default timeout value to ImplementationValidator itself

* Increase default connection timeout to 3+epsilon for TCP retransmission

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.nore
[…]
