### 2021-07-06 a10e6bb
Bump OPTIMADE version and fix README
