### 2021-07-06 5e3e45e
Add issue_tracker field to provider model - Update OPTIMADE API to v1.1 (#593)
Added 'issue_tracker' field from
https://github.com/Materials-Consortia/OPTIMADE/pull/339
Add default values for issue_tracker to configuration files.

Bump OPTIMADE version to v1.1.0 and fix README.
