+++ optimade/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.15.5"
+__version__ = "0.16.0"
 __api_version__ = "1.1.0"
