### 2021-07-07 #3458 @whedon
[REVIEW]: optimade-python-tools: a Python library for serving and consuming materials data via OPTIMADE APIs
**Submitting author:** @ml-evs (<a href="http://orcid.org/0000-0002-1182-9098">Matthew L Evans</a>)
**Repository:** <a href="https://github.com/Materials-Consortia/optimade-python-tools" target ="_blank">https://github.com/Materials-Consortia/optimade-python-tools</a>
**Version:** 0.16.4

- [x] #3542
