### 2021-07-07 #3458 @ml-evs
FYI @jgostick, I've just been made aware that James Hester (@jamesrhester) is a collaborator on an ongoing project with one of the authors of this package (@CasperWA) and some of the authors of the OPTIMADE specification more widely.
