### 2021-07-08 #3458 @jamesrhester
@jgostick, I have read the COI policy and should declare here that author @CasperWA and myself are both currently involved in the [EMMO crystallography project](https://github.com/emmo-repo/CIF-ontology/). I have also participated in several zoom meetings of the OPTIMADE consortium, neither of which concerned this work, of which I was not aware until now. Neither my institution nor myself have received any funding for this work and I believe that I can assess this paper objectively.

Please advise if JOSS is prepared to waive any actual or perceived conflicts of interest in this case.
