### 2021-07-09 2638bac
Update description of element names for v1.1
