### 2021-07-12 849cc76
Some additional tests for HAS ONLY filters
