### 2021-07-12 7b9c683
Convert headers warning to pure Python warning, not validator warning
