### 2021-07-12 7725f12
Remove unnecessary else after while/return
