### 2021-07-12 a79ce79
Apply suggestions from code review
Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>
