### 2021-07-13 1204981
Minor type and documentation corrections
