+++ optimade/server/routers/utils.py
@@ -1,8 +1,8 @@
 # pylint: disable=import-outside-toplevel,too-many-locals
 import re
-import urllib
+import urllib.parse
 from datetime import datetime
-from typing import Union, List, Dict, Set
+from typing import Any, Dict, List, Set, Union
 
 from fastapi import HTTPException, Request
 from starlette.datastructures import URL as StarletteURL
@@ -66,7 +66,7 @@ def handle_response_fields(
     results: Union[List[EntryResource], EntryResource],
     exclude_fields: Set[str],
     include_fields: Set[str],
-) -> List[dict]:
+) -> List[Dict[str, Any]]:
     """Handle query parameter `response_fields`.
 
     It is assumed that all fields are under `attributes`.
