### 2021-07-13 bc29d08
Correct the behaviour of HAS ONLY with MongoDB backend (#861)
* Corrected the behaviour of HAS ONLY and updated tests with the correct results.

* Rename _apply_only_filter to _apply_has_only_filter

Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>

* Gave "structures.id" the same behaviour as "references.id" for HAS_ONLY filters."

* Some additional tests for HAS ONLY filters

* Apply suggestions from code review

Co-authored-by: Matthew Evans <7916000+ml-evs@users.noreply.github.com>

* Added import of Badrequest from optimade.server.exceptions.

Co-authored-by: Matthe
[…]
