### 2021-07-14 90ad30b
Minor corrections + use model aliases for `handle_response_fields()` (#876)
Minor type and documentation corrections.

Return results by alias when handling `response_fields`.
This will use the pydantic parameter `by_alias=True` when converting
models to dictionaries in the `handle_response_fields()` utility func.
