+++ optimade/server/entry_collections/entry_collections.py
@@ -4,12 +4,11 @@ import warnings
 import re
 
 from lark import Transformer
-from fastapi import HTTPException
 
 from optimade.filterparser import LarkParser
 from optimade.models import EntryResource
 from optimade.server.config import CONFIG, SupportedBackend
-from optimade.server.exceptions import BadRequest, Forbidden
+from optimade.server.exceptions import BadRequest, Forbidden, NotFound
 from optimade.server.mappers import BaseResourceMapper
 from optimade.server.query_params import EntryListingQueryParams, SingleEntryQueryParams
 from optimade.server.warnings import FieldValueNotRecognized, UnknownProviderProperty
@@ -86,6 +85,8 @@ class EntryCollection(ABC):
         self.provider_prefix = CONFIG.provider.prefix
         self.provider_fields = CONFIG.provider_fields.get(resource_mapper.ENDPOINT, [])
 
+        self._all_fields: Set[str] = None
+
     @abstractmethod
     def __len__(self) -> int:
         """Returns the total number of entries in the collection."""
@@ -110,8 +111,10 @@ class EntryCollection(ABC):
         """
 
     def find(
-        self, params: EntryListingQueryParams
-    ) -> Tuple[List[EntryResource], int, bool, set]:
+        self, params: Union[EntryListingQueryParams, SingleEntryQueryParams]
+    ) -> Tuple[
+        Union[List[EntryResource], EntryResource, None], int, bool, Set[str], Set[str]
+    ]:
         """
         Fetches results and indicates if more data is available.
 
@@ -120,10 +123,11 @@ class EntryCollection(ABC):
         for more information.
 
         Parameters:
-            params (EntryListingQueryParams): entry listing URL query params
+            params: Entry listing URL query params.
 
         Returns:
-            (`results`, `data_returned`, `more_data_available`, `fields`).
+            A tuple of various relevant values:
+            (`results`, `data_returned`, `more_data_available`, `exclude_fields`, `include_fields`).
 
         """
         criteria = self.handle_query_params(params)
@@ -131,15 +135,14 @@ class EntryCollection(ABC):
         response_fields = criteria.pop("fields")
 
         results, data_returned, more_data_available = self._run_db_query(
-            criteria, single_entry=isinstance(params, SingleEntryQueryParams)
+            criteria, single_entry
         )
 
         if single_entry:
             results = results[0] if results else None
 
             if data_returned > 1:
-                raise HTTPException(
-                    status_code=404,
+                raise NotFound(
                     detail=f"Instead of a single entry, {data_returned} entries were found",
                 )
 
@@ -200,24 +203,30 @@ class EntryCollection(ABC):
         """
 
     @property
-    def all_fields(self) -> set:
+    def all_fields(self) -> Set[str]:
         """Get the set of all fields handled in this collection,
         from attribute fields in the schema, provider fields and top-level OPTIMADE fields.
+
+        The set of all fields are lazily created and then cached.
+        This means the set is created the first time the property is requested and then cached.
 
         Returns:
             All fields handled in this collection.
 
         """
-        # All OPTIMADE fields
-        fields = self.resource_mapper.TOP_LEVEL_NON_ATTRIBUTES_FIELDS.copy()
-        fields |= self.get_attribute_fields()
-        # All provider-specific fields
-        fields |= {
-            f"_{self.provider_prefix}_{field_name}"
-            for field_name in self.provider_fields
-        }
+        if not self._all_fields:
+            # All OPTIMADE fields
+            self._all_fields = (
+                self.resource_mapper.TOP_LEVEL_NON_ATTRIBUTES_FIELDS.copy()
+            )
[…]
