### 2021-07-14 7986cac
Change default MIME type to 'application/vnd.api+json'
- This is to follow the JSON:API spec requirement on Content Negotiation
  https://jsonapi.org/format/#content-negotiation.
