### 2021-07-14 e04b237
Replace exception Mixin with new base class
