+++ optimade/server/exceptions.py
@@ -1,4 +1,5 @@
-from fastapi import HTTPException
+from abc import ABC
+from fastapi import HTTPException as FastAPIHTTPException
 
 __all__ = (
     "BadRequest",
@@ -11,112 +12,70 @@ __all__ = (
 )
 
 
-class StrReprMixin(HTTPException):
-    """This mixin can be useful when testing requires a string
-    representation of an exception that contains the HTTPException
+class HTTPException(FastAPIHTTPException, ABC):
+    """This abstract class makes it easier to subclass FastAPI's HTTPException with
+    new status codes.
+
+    It can also be useful when testing requires a string representation
+    of an exception that contains the HTTPException
     detail string, rather than the standard Python exception message.
+
     """
+
+    def __init__(self, detail: str = None, headers: dict = None) -> None:
+        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
     def __str__(self):
         return self.detail if self.detail is not None else self.__repr__()
 
 
-class BadRequest(StrReprMixin, HTTPException):
+class BadRequest(HTTPException):
     """400 Bad Request"""
 
     status_code: int = 400
     title: str = "Bad Request"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class VersionNotSupported(StrReprMixin, HTTPException):
+class VersionNotSupported(HTTPException):
     """553 Version Not Supported"""
 
     status_code: int = 553
     title: str = "Version Not Supported"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class Forbidden(StrReprMixin, HTTPException):
+class Forbidden(HTTPException):
     """403 Forbidden"""
 
     status_code: int = 403
     title: str = "Forbidden"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class NotFound(StrReprMixin, HTTPException):
+class NotFound(HTTPException):
     """404 Not Found"""
 
     status_code: int = 404
     title: str = "Not Found"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class UnprocessableEntity(StrReprMixin, HTTPException):
+class UnprocessableEntity(HTTPException):
     """422 Unprocessable Entity"""
 
     status_code: int = 422
     title: str = "Unprocessable Entity"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class InternalServerError(StrReprMixin, HTTPException):
+class InternalServerError(HTTPException):
     """500 Internal Server Error"""
 
     status_code: int = 500
     title: str = "Internal Server Error"
 
-    def __init__(
-        self,
-        detail: str = None,
-        headers: dict = None,
-    ) -> None:
-        super().__init__(status_code=self.status_code, detail=detail, headers=headers)
 
-
-class NotImplementedResponse(StrReprMixin, HTTPException):
+class NotImplementedResponse(HTTPException):
     """501 Not Implemented"""
 
     status_code: int = 501
     title: str = "Not Implemented"
-
-    def __init__(
-        self,
[…]
