### 2021-07-14 52d4f4f
Use specific defined HTTP exceptions everywhere
