### 2021-07-14 06f45be
Tweaks from code review
