### 2021-07-15 c9975aa
Add missing __str__ type hint
Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
