### 2021-07-15 e4c07cc
Change default MIME type to "application/vnd.api+json" (#877)
* Change default MIME type to 'application/vnd.api+json'

- This is to follow the JSON:API spec requirement on Content Negotiation
  https://jsonapi.org/format/#content-negotiation.

* Replace exception Mixin with new base class

* Use specific defined HTTP exceptions everywhere

* Tweaks from code review

* Add missing __str__ type hint

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>
