+++ optimade/server/main_index.py
@@ -22,7 +22,7 @@ from optimade.server.logger import LOGGER
 from optimade.server.exception_handlers import OPTIMADE_EXCEPTIONS
 from optimade.server.middleware import OPTIMADE_MIDDLEWARE
 from optimade.server.routers import index_info, links, versions
-from optimade.server.routers.utils import BASE_URL_PREFIXES
+from optimade.server.routers.utils import BASE_URL_PREFIXES, JSONAPIResponse
 
 if os.getenv("OPTIMADE_CONFIG_FILE") is None:
     LOGGER.warn(
@@ -49,6 +49,7 @@ This specification is generated using [`optimade-python-tools`](https://github.c
     docs_url=f"{BASE_URL_PREFIXES['major']}/extensions/docs",
     redoc_url=f"{BASE_URL_PREFIXES['major']}/extensions/redoc",
     openapi_url=f"{BASE_URL_PREFIXES['major']}/extensions/openapi.json",
+    default_response_class=JSONAPIResponse,
 )
 
 
