+++ tests/server/conftest.py
@@ -75,6 +75,10 @@ def get_good_response(client, index_client):
             response = used_client.get(request, **kwargs)
             response_json = response.json()
             assert response.status_code == 200, f"Request failed: {response_json}"
+            expected_mime_type = "application/vnd.api+json"
+            assert (
+                response.headers["content-type"] == expected_mime_type
+            ), f"Response should have MIME type {expected_mime_type!r}, not {response.headers['content-type']!r}."
         except json.JSONDecodeError:
             print(
                 f"Request attempted:\n{used_client.base_url}{used_client.version}"
@@ -174,6 +178,10 @@ def check_error_response(client, index_client):
                 f"Request should have been an error with status code {expected_status}, "
                 f"but instead {response.status_code} was received.\nResponse:\n{response.json()}",
             )
+            expected_mime_type = "application/vnd.api+json"
+            assert (
+                response.headers["content-type"] == expected_mime_type
+            ), f"Response should have MIME type {expected_mime_type!r}, not {response.headers['content-type']!r}."
 
             response = response.json()
             assert len(response["errors"]) == 1, response.get(
