### 2021-07-27 #3458 @danielskatz
Given that we have to balance the needed expertise to review a submission and the fact that this often means that potential reviewers will be involved in community projects with the developers, this needs to be judged on a case-by-case basis. I think that, given @jamesrhester's comments, we can waive the COI and proceed, if that seems ok with the editor of this submission, @jgostick
