### 2021-07-27 4c2d2a5
Consolidated use case docs, added NOMAD and reworded integrated tutorial
