+++ docs/getting_started/integrated.md
@@ -1,35 +1,33 @@
 # Integrate OPTIMADE with an existing web application
 
-`optimade-python-tools` (OPT) can be run as a standalone web application that serves the OPTIMADE API based on a pre-configured MongoDB backend.
-In this document, we are going to use OPT differently and use it to add an OPTIMADE API implementation alongside an existing API employing an Elasticsearch storage layer.
+The `optimade` package can be used to create a standalone web application that serves the OPTIMADE API based on a pre-configured MongoDB backend.
+In this document, we are going to use `optimade` differently and use it to add an OPTIMADE API implementation alongside an existing API that employs an Elasticsearch storage layer.
 
 Let's assume we already have a *FastAPI* application that runs an unrelated web service, and that we use an Elasticsearch backend that contains all structure data, but not necessarily in a form that OPTIMADE expects.
 
-## Providing an `optimade-python-tools` configuration
+## Providing the `optimade` configuration
 
-OPT can read its configuration from a JSON file.
+`optimade` can read its configuration from a JSON file.
 It uses the `OPTIMADE_CONFIG_FILE` environment variable (or a default path) to find the config file.
-If you run OPT from another application, you might want to provide this config file as part of the source code and not via environment variables.
-Let's say you have a file `optimade_config.json` as part of the Python module that you use to create your OPT.
+If you run `optimade` code inside another application, you might want to provide this config file as part of the source code and not via environment variables.
+Let's say you have a file `optimade_config.json` as part of the Python module that you use to create your OPTIMADE API.
 
 !!! tip
-    You can find more detailed information about configuring OPT in the [Configuration](../configuration.md) section.
+    You can find more detailed information about configuring the `optimade` server in the [Configuration](../configuration.md) section.
 
-Before importing any OPT packages, you can set the `OPTIMADE_CONFIG_FILE` environment variable to link OPT to your config file:
+Before importing any `optimade` modules, you can set the `OPTIMADE_CONFIG_FILE` environment variable to refer to your config file:
 
 ```python
 import os
-import sys
+from pathlib import Path
 
-os.environ['OPTIMADE_CONFIG_FILE'] = os.path.join(
-    os.path.dirname(__file__), 'optimade_config.json'
-)
+os.environ['OPTIMADE_CONFIG_FILE'] = str(Path(__file__).parent / "optimade_config.json")
 ```
 
 ## Customize the [`EntryCollection`][optimade.server.entry_collections.entry_collections.EntryCollection] implementation
 
-Let's assume that your Elasticsearch backend stores structure data different enough that you need to provide your own custom implementation.
-The following code customizes the used [`EntryCollection`][optimade.server.entry_collections.entry_collections.EntryCollection] class for structures, while keeping the default *mongomock* implementation (using [`MongoCollection`][optimade.server.entry_collections.mongo.MongoCollection]) for all other entry types.
+Let's assume that your Elasticsearch backend stores structure data in a different enough manner that you need to provide your own custom implementation.
+The following code customizes the [`EntryCollection`][optimade.server.entry_collections.entry_collections.EntryCollection] class for structures, whilst keeping the default MongoDB-based implementation (using [`MongoCollection`][optimade.server.entry_collections.mongo.MongoCollection]) for all other entry types.
 
 ```python
 from optimade.server.routers import structures
@@ -37,19 +35,20 @@ from optimade.server.routers import structures
 structures.structures_coll = MyElasticsearchStructureCollection()
 ```
 
[…]
