### 2021-07-27 7d25322
Consolidated use case docs, added NOMAD and reworded integrated tutorial
