+++ requirements.txt
@@ -2,7 +2,7 @@ fastapi==0.65.2
 lark-parser==0.11.3
 pydantic==1.8.2
 email_validator==1.1.3
-requests==2.25.1
+requests==2.26.0
 uvicorn==0.14.0
 pymongo==3.11.4
 mongomock==3.23.0
