### 2021-08-01 11d764d
Bump elasticsearch-dsl from 6.4.0 to 7.4.0
Bumps [elasticsearch-dsl](https://github.com/elasticsearch/elasticsearch-dsl-py) from 6.4.0 to 7.4.0.
- [Release notes](https://github.com/elasticsearch/elasticsearch-dsl-py/releases)
- [Changelog](https://github.com/elastic/elasticsearch-dsl-py/blob/master/Changelog.rst)
- [Commits](https://github.com/elasticsearch/elasticsearch-dsl-py/compare/6.4.0...v7.4.0)

---
updated-dependencies:
- dependency-name: elasticsearch-dsl
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@githu
[…]
