+++ requirements.txt
@@ -6,7 +6,7 @@ requests==2.25.1
 uvicorn==0.14.0
 pymongo==3.11.4
 mongomock==3.23.0
-elasticsearch-dsl==6.4.0
+elasticsearch-dsl==7.4.0
 Jinja2==3.0.1
 typing-extensions==3.10.0.0
 pyyaml==5.4.1
