### 2021-08-01 d17193d
Bump jarvis-tools from 2021.6.18 to 2021.7.19
Bumps [jarvis-tools](https://github.com/usnistgov/jarvis) from 2021.6.18 to 2021.7.19.
- [Release notes](https://github.com/usnistgov/jarvis/releases)
- [Commits](https://github.com/usnistgov/jarvis/compare/v2021.06.18...v2021.07.19)

---
updated-dependencies:
- dependency-name: jarvis-tools
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
