### 2021-08-05 #3458 @jgostick
It seems I am not being notified by github of mentions.  I'm very sorry for taking so long to check in, I just do it once a week or so.

Regarding the COI, I would say that the more ably a person can review a submission the closer they are to the COI threshold.  I'm fine with things proceeding as they are.
