### 2021-08-06 fd78ce2
Fix provider fallbacks (#897)
* Fix provider list fallback

* Make the provider list importable
