### 2021-08-06 f7bdea9
Test whether we can get away with depless AiiDA install
