+++ requirements-client.txt
@@ -1,4 +1,4 @@
-aiida-core==1.6.4
+aiida-core==1.6.4 --install-option=no-deps
 ase==3.22.0
 numpy==1.21.0
 pymatgen==2021.3.9
