### 2021-08-06 f79c9b8
Dependency updates for v0.16.2 (#894)
* Bump pymongo from 3.11.4 to 3.12.0 (#892)

Bumps [pymongo](https://github.com/mongodb/mongo-python-driver) from 3.11.4 to 3.12.0.
- [Release notes](https://github.com/mongodb/mongo-python-driver/releases)
- [Changelog](https://github.com/mongodb/mongo-python-driver/blob/3.12.0/doc/changelog.rst)
- [Commits](https://github.com/mongodb/mongo-python-driver/compare/3.11.4...3.12.0)

---
updated-dependencies:
- dependency-name: pymongo
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@
[…]
