+++ requirements-docs.txt
@@ -1,4 +1,4 @@
-mkdocs==1.2.1
+mkdocs==1.2.2
 mkdocs-awesome-pages-plugin==2.5.0
 mkdocs-material==7.2.2
 mkdocs-minify-plugin==0.4.0
