### 2021-08-06 db171a7
Bump fastapi from 0.65.2 to 0.68.0
Bumps [fastapi](https://github.com/tiangolo/fastapi) from 0.65.2 to 0.68.0.
- [Release notes](https://github.com/tiangolo/fastapi/releases)
- [Commits](https://github.com/tiangolo/fastapi/compare/0.65.2...0.68.0)

---
updated-dependencies:
- dependency-name: fastapi
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
