+++ requirements.txt
@@ -1,4 +1,4 @@
-fastapi==0.65.2
+fastapi==0.68.0
 lark-parser==0.11.3
 pydantic==1.8.2
 email_validator==1.1.3
