### 2021-08-10 bd59b5c
Relax response top-level root validator
Add test that checks a `Response` can be initialized with "null" values
for the "required" top-level fields, while still raising a
`ValidationError` if the "required" top-level fields are left out.
