### 2021-08-10 34772b7
Relax validations for general response models
This related mainly to the top-level fields and their presence or not,
instead of validating their content, which is too strict.
The exception being the top-level field `errors`.

Also added a lot of EOL periods to descriptions and doc-strings.
