### 2021-08-11 1d951bf
Relax response top-level root validator (#903)
Add test that checks a `Response` can be initialized with "null" values
for the "required" top-level fields, while still raising a
`ValidationError` if the "required" top-level fields are left out.

Relax validations for general response models.
This related mainly to the top-level fields and their presence or not,
instead of validating their content, which is too strict.
The exception being the top-level field `errors`.

Also added a lot of EOL periods to descriptions and doc-strings.
