### 2021-08-20 #3458 @jamesrhester
A few queries and suggestions below - the remainder I'll raise as issues in the repository as they don't concern the paper itself:

1. Github "contributors" suggests that Michael Wu has contributed significantly more to the code than some of the co-authors. Is there any reason he is not a co-author (I see he is acknowledged)?
2. The Lark reference points to pypi, but the other code references are to Github. Is there any reason the Lark pypi link is preferred over Github?
3. There is no mention in the paper of any alternatives to optimade-python-tools. If there are none, a sentence to this effect could be added, e.g. "The authors are not aware of any other packages with similar functionality".
3. The paper could be made appropriate to a broader audience if a few small
clarifications are included as follows:

- The term "API" can include APIs for programming languages. While it
eventually becomes clear to the reader that web APIs are the particular
intended meaning, I suggest that in the "Summary" when the term "API"
is first introduced, clarify that a standard for web-based access is
being discussed, e.g.

 "Access to these databases has been made available through bespoke web-based
 protocols, commonly termed Application Programming Interfaces (APIs), typically
 implemented through specially-constructed stateless HTTP queries."

- Likewise, the term "REST" API is used without definition. I suggest adding
"(stateless)" after the first use of the acronym "REST".
