### 2021-08-20 22ae09c
Clarify installation methods by use case
