### 2021-08-21 #3458 @ml-evs
Thanks for your initial comments @jamesrhester. We can handle any further issues directly in the repository, but I have addressed your suggested changes to the paper below. To answer your comments in reverse order:

4. I agree with your comments, and have tried to adopt your suggestions in such a way that knowledge of the requisite jargon isn't a necessary (I hope URL and HTTP can be taken as assumed knowledge, as constructing a legible sentence with so many acronyms is a challenge!):
      - In the summary, the sentence in question has been made more specific:
        ```diff
        - These databases have been made available with bespoke application programming interfaces (APIs) to allow for automated and often open access to the underlying data.
        + Each database will typically have a bespoke, stateless, web-based Application Programming Interface (API); a users can submit a query via a specially-crafted URLs (which encodes a more general HTTP request).
        ```
      - I have removed the term "REST" from the sentence in the statement of need (it remains in the keywords for the paper).
3. I have added the following sentence to the statement of need:
    ```diff
    + As this package was developed concomitantly with the OPTIMADE specification itself, the present authors are not aware of any other generic packages with similar functionality.
    ```
2. Thank you for spotting this, the Lark reference now points to the appropriate GitHub repo (unfortunately none of the software packages we cite provide a recommended citation or Zenodo archive).
1. Thank you for raising this - we agree that appropriate credit is very important for this kind of work. Michael's useful contributions to the repository were at an early exploratory stage before it took it's current form (the final section of the automated software report above provides some indication of this). Those present on the author list of the paper have also made non-code contributions in terms of design and feedback (as the package's development was guided by their OPTIMADE implementations). I should also note that all of the contributors to the package (prior to the paper submission, at least) were present as authors to the OPTIMADE specification paper, so are credited there for any work they performed in the early design stages of the overall endeavor. It does not seem to be typical for JOSS papers to include contribution statements, but we would be happy to add one.
