### 2021-08-21 9847bf4
Clarify installation methods by use-case (#905)
