### 2021-08-21 dae5d07
Regenerate OpenAPI schemas with FastAPI 0.68.0
