### 2021-08-21 461ae1d
Make 'sortable', 'support', 'queryable' use the 'x-optimade-' namespace
In FastAPI 0.68.0, any keys passed through to the pydantic Field
are added to the OpenAPI schema, without any prefix. This meant
that the constructed OpenAPI schemas were invalid, so we have to
manually patch the custom OPTIMADE keys so that they get added
with the 'x-optimade' prefix.
