### 2021-08-26 eb0267f
Add more specific 'Getting Help' info to contributing and readme
