### 2021-08-26 a110c53
Add some more blurb around the model documentation
