### 2021-08-26 ecf26d3
Add citation data and remove 'status' section from README
