### 2021-08-26 7f731c4
Add more specific 'Getting Help' info and restructure README (#911)
* Add more specific 'Getting Help' info to contributing and readme

* Minor rewording

* Add citation data and remove 'status' section from README
