### 2021-08-26 7a8e529
Clarify the "all models" documentation page (#912)
* Add some more blurb around the model documentation

* Tweak wording introducing all_models page

Co-authored-by: Johan Bergsma <JPBergsma@users.noreply.github.com>
