### 2021-08-27 4adbaa6
Minor tweaks to the validator
- Add --version flag for the validator
- Do 'minimal' tests first, as this is where actual model deserialization occurs
