+++ requirements.txt
@@ -3,7 +3,7 @@ lark-parser==0.11.3
 pydantic==1.8.2
 email_validator==1.1.3
 requests==2.26.0
-uvicorn==0.14.0
+uvicorn==0.15.0
 pymongo==3.12.0
 mongomock==3.23.0
 elasticsearch-dsl==6.4.0
