### 2021-08-30 d648fe7
Bump typing-extensions from 3.10.0.0 to 3.10.0.1
Bumps [typing-extensions](https://github.com/python/typing) from 3.10.0.0 to 3.10.0.1.
- [Release notes](https://github.com/python/typing/releases)
- [Commits](https://github.com/python/typing/compare/3.10.0.0...3.10.0.1)

---
updated-dependencies:
- dependency-name: typing-extensions
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
