+++ requirements-docs.txt
@@ -1,5 +1,5 @@
 mkdocs==1.2.2
 mkdocs-awesome-pages-plugin==2.5.0
-mkdocs-material==7.2.2
+mkdocs-material==7.2.5
 mkdocs-minify-plugin==0.4.0
 mkdocstrings==0.15.2
