### 2021-08-30 0e358b2
Bump jarvis-tools from 2021.7.19 to 2021.8.18
Bumps [jarvis-tools](https://github.com/usnistgov/jarvis) from 2021.7.19 to 2021.8.18.
- [Release notes](https://github.com/usnistgov/jarvis/releases)
- [Commits](https://github.com/usnistgov/jarvis/compare/v2021.07.19...v2021.08.18)

---
updated-dependencies:
- dependency-name: jarvis-tools
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
