### 2021-08-30 0798e9a
Bump lark-parser from 0.11.3 to 0.12.0
Bumps [lark-parser](https://github.com/lark-parser/lark) from 0.11.3 to 0.12.0.
- [Release notes](https://github.com/lark-parser/lark/releases)
- [Commits](https://github.com/lark-parser/lark/compare/0.11.3...0.12.0)

---
updated-dependencies:
- dependency-name: lark-parser
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
