+++ requirements.txt
@@ -1,5 +1,5 @@
 fastapi==0.65.2
-lark-parser==0.11.3
+lark-parser==0.12.0
 pydantic==1.8.2
 email_validator==1.1.3
 requests==2.26.0
