### 2021-08-30 00f0e3b
Bump numpy from 1.21.1 to 1.21.2
Bumps [numpy](https://github.com/numpy/numpy) from 1.21.1 to 1.21.2.
- [Release notes](https://github.com/numpy/numpy/releases)
- [Changelog](https://github.com/numpy/numpy/blob/main/doc/HOWTO_RELEASE.rst.txt)
- [Commits](https://github.com/numpy/numpy/compare/v1.21.1...v1.21.2)

---
updated-dependencies:
- dependency-name: numpy
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
