+++ requirements-client.txt
@@ -1,5 +1,5 @@
 aiida-core==1.6.5
 ase==3.22.0
-numpy==1.21.1
+numpy==1.21.2
 pymatgen==2021.3.9
 jarvis-tools==2021.8.18
