### 2021-09-01 52420ba
Bump typing-extensions from 3.10.0.1 to 3.10.0.2
Bumps [typing-extensions](https://github.com/python/typing) from 3.10.0.1 to 3.10.0.2.
- [Release notes](https://github.com/python/typing/releases)
- [Commits](https://github.com/python/typing/compare/3.10.0.1...3.10.0.2)

---
updated-dependencies:
- dependency-name: typing-extensions
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
