+++ requirements.txt
@@ -8,5 +8,5 @@ pymongo==3.12.0
 mongomock==3.23.0
 elasticsearch-dsl==6.4.0
 Jinja2==3.0.1
-typing-extensions==3.10.0.1
+typing-extensions==3.10.0.2
 pyyaml==5.4.1
