### 2021-09-01 79d932a
Fix config file warnings when specified file is missing (#931)
