### 2021-09-01 2ad3199
Add end-to-end guide for setting up an OPTIMADE API
