+++ docs/getting_started/setting_up_an_api.md
@@ -1,0 +1,112 @@
+# Setting up an OPTIMADE API
+
+These notes describe how to set up and customize an OPTIMADE API based on the reference server in this package for some existing crystal structure data.
+
+To follow this guide, you will need to have a working development installation, as described in the [installation instructions](../INSTALL.md#full-development-installation).
+Complete examples of APIs that use this package are described in the [Use Cases](./use_cases.md) section.
+
+## Setting up the database
+
+The `optimade` reference server requires a data source per OPTIMADE entry type (`structures`, `references`, `links`).
+In the simplest case, these can be configured as named MongoDB collections with a defined MongoDB URI and database name (see below), but they can also be set up as custom subclasses of [`EntryCollection`][optimade.server.entry_collections.entry_collections.EntryCollection] that could simply read from a static file.
+In the reference server, these data sources, or collections, are created in the submodule for the corresponding routers/endpoints.
+
+Here, we shall use the built-in MongoDB collections for each entry type, by simply specifying the appropriate options in the [configuration](../configuration.md), namely [`"database_backend": "mongodb"`][optimade.server.config.ServerConfig.database_backend], [`"mongo_uri": "mongodb://localhost:27017"`][optimade.server.config.ServerConfig.mongo_uri], [`"mongo_database": "optimade"`][optimade.server.config.ServerConfig.mongo_database] and the collection names for each entry type ([`"structures_collection": "structures"`][optimade.server.config.ServerConfig.structures_collection] etc.).
+These notes will now assume that you have a MongoDB instance running and you have created a database that matches your [`"mongo_database"`][optimade.server.config.ServerConfig.mongo_database] config option.
+
+If you disable inserting test data (with the [`"insert_test_data": false`][optimade.server.config.ServerConfig.insert_test_data] configuration option), you can test your API/database connection by running the web server with `uvicorn optimade.main:app` and visiting the (hopefully empty) structures endpoint at `localhost:5000/v1/structures` (or your chosen base URL).
+
+!!! note
+    As of version v0.16, the other supported database backend is Elasticsearch.
+    If you are interested in using another backend, or would like it to be supported in the `optimade` package, please raise an issue on [GitHub](https://github.com/Materials-Consortia/optimade-python-tools/issues/new) and visit the notes on implementing new [filter transformers](./filtering.md#developing-new-filter-transformers).
+
+## Mapping non-OPTIMADE data
+
+There are two ways to work with data that does not exactly match the OPTIMADE specification, both of which require configuring a subclass of [`BaseResourceMapper`][optimade.server.mappers.entries.BaseResourceMapper] that converts your stored data format into an OPTIMADE-compliant entry.
+The two options are:
+
+- Use the mapper to dynamically convert the data stored in the database, and the filters on that data, to an OPTIMADE format when responding to API requests.
+- Apply the mapper to your entries before ingestion and use it to create a secondary database that stores the converted entries (e.g., normalized data), or equivalently, adding all the required OPTIMADE fields inside the existing entries (e.g., denormalized data)
+
+The main consideration when choosing these options is not necessarily how closely your data matches the OPTIMADE format, but instead how readily the OPTIMADE filtering of that document can be mapped into the corresponding database query.
+This could require writing or extending the [`BaseFilterTransformer`][optimade.filtertransformers.base_transformer.BaseTransformer] class, which takes an OPTIMADE filter string and converts it into a backend-specific query.
+
[…]
