### 2021-09-01 226fb9b
Add some docstrings and type-hints for importable constants/singletons
