### 2021-09-02 2217a63
Bump deps in setup.py
