### 2021-09-02 06cf0ed
Use proper pip dependency resolver (closes \#625)
