### 2021-09-07 b7becc7
remove the dependency on mkdocs-minify because of issue #938.
