### 2021-09-09 #3458 @hungpham2017
Dear all,

I apologize for being too late. Overall, I find that this is a great project. I have tried to use the package myself as rather an inexperienced user to see how easy one could start with no previous experience. Here are some minor comments/suggestions that mostly I hope they would make the projects more popular to the community:

(1) Apparently, there is an old (v0.16.3) release of optimade on conda-forge. The latest version of pip is v1.0.0. Is there any reason why the latest version is not there? It could reach far more user if it is available there.

(2) Although the application of optimade-python-tools on the Materials Cloud was mentioned in the paper. Its use via [https://www.materialscloud.org/work/tools/optimadeclient](url) wasn't. At least to me, this is a great example to show how OPTIMADE is applied to many other open-source project. The Software paper would be useful is this is mentioned.

(3) To me, the documentations are quite scattered since there is a few different places ([https://github.com/Materials-Consortia/optimade-python-tools](url) , [https://www.optimade.org/optimade-python-tools/getting_started/use_cases/](url), [https://www.optimade.org/](url) . Although they are all informative, it would be easier to stay focus on one single document page so that the new users can quickly narrow down where they need to look at, let's say the installation instruction or a tutorial. I needed to be patient to check around different resources for OPTIMADE and optimize-python-tools to get familiar with the project.

(4) I may be missing something. Most of the examples on [https://www.optimade.org/optimade-python-tools/](url) are for the developers who would like to implement OPMADE on their web applications. Is it possible for users to inquire a material information (e.g., lattice constants, band gap) from a database of choice via a simple python script without interacting with the web service. I don't even know this makes sense ... I couldn't find something close to this on the Getting started page

Rather than the above suggestions, I would be happy to recommend this project/paper to be accepted.

Best,
Hung
