### 2021-09-10 #3458 @CasperWA
Dear Hung,

> I apologize for being too late. Overall, I find that this is a great project. I have tried to use the package myself as rather an inexperienced user to see how easy one could start with no previous experience (...)

This seems like it could be a great stress test, indeed 👍
And thank you for taking the time to do this! I'll try to respond to your points as best I can.

> (1) Apparently, there is an old (v0.16.3) release of optimade on conda-forge. The latest version of pip is v1.0.0. Is there any reason why the latest version is not there? It could reach far more user if it is available there.

`v0.16.3` is currently the latest version.
I am unsure where you found a v1.0.0?
The conda-forge release is here: https://anaconda.org/conda-forge/optimade
The PyPI release is here: https://pypi.org/project/optimade/
And it should all match up with the GitHub releases as well, which is here: https://github.com/Materials-Consortia/optimade-python-tools/releases

> (2) Although the application of optimade-python-tools on the Materials Cloud was mentioned in the paper. Its use via [https://www.materialscloud.org/work/tools/optimadeclient](url) wasn't. At least to me, this is a great example to show how OPTIMADE is applied to many other open-source project. The Software paper would be useful is this is mentioned.

Agreed! This could easily be added to the "Use cases" section with a focus on the `optimade.adapters` module.

> (3) To me, the documentations are quite scattered since there is a few different places ([https://github.com/Materials-Consortia/optimade-python-tools](url) , [https://www.optimade.org/optimade-python-tools/getting_started/use_cases/](url), [https://www.optimade.org/](url) . Although they are all informative, it would be easier to stay focus on one single document page so that the new users can quickly narrow down where they need to look at, let's say the installation instruction or a tutorial. I needed to be patient to check around different resources for OPTIMADE and optimize-python-tools to get familiar with the project.

Good comment.
The official place with documentation on optimade-python-tools is at https://www.optimade.org/optimade-python-tools.

We will try to make this clearer on the GitHub README page, as well as discuss with the OPTIMADE consortium to make this clearer on the central OPTIMADE website (https://www.optimade.org), which caters and serves information about the OPTIMADE consortium in general and not just the Python tools package.

> (4) I may be missing something. Most of the examples on [https://www.optimade.org/optimade-python-tools/](url) are for the developers who would like to implement OPMADE on their web applications. Is it possible for users to inquire a material information (e.g., lattice constants, band gap) from a database of choice via a simple python script without interacting with the web service. I don't even know this makes sense ... I couldn't find something close to this on the Getting started page

I am not completely sure if I understood this correctly.
But I'll try to respond as I understand it :)

By "web service" I am assuming you mean one of the online OPTIMADE clients that are available (an overview can be seen on the web site here: https://www.optimade.org/clients)? Note, there is also an in-development client available through the `pymatgen` Python package.
If this is correctly assumed, then it is both true that the Python package `optimade` and the GitHub repository `optimade-python-tools` do not have a fully-fledged or even in-development client and that one can write a simple Python script to retrieve materials science data information from one of the OPTIMADE database providers - either utilizing the `optimade` Python package for enhanced tools or not.
The enhanced tools here specifically being the `optimade.adapters` module, with which one can instantiate each of the JSON objects returned under the top-level `data` list of an OPTIMADE response with a Python class from the `optimade.adapters` module, e.g., `optimade.adapters.Structure`. This class will utilize the `pydantic` data models from the `optimade.models` module to instantiate and validate the data structure and all `optimade.adapters` resource classes also have methods to convert the OPTIMADE datum to other meaningful formats, depending on the packages you have installed in your Python environment. For material structures this includes CIF, PDB, PDBx/mmCIF if NumPy is installed, these can all be written to a file and stored. It can also be converted to other Python packages' materials structure representations: `ase.Atoms` if ASE is installed, `pymatgen.core.Structure`/`pymatgen.core.Molecule` if pymatgen is installed, `jarvis.core.atoms.Atoms` if JARVIS-Tools is installed, and `aiida.orm.StructureData` if AiiDA is installed.

However, to perform the actual OPTIMADE queries, one would need to use another package that supports HTTP requests (like `requests` or the built-in `urllib`) and follow the OPTIMADE specification (https://github.com/Materials-Consortia/OPTIMADE/blob/master/optimade.rst) on how to create the URL for the specific query you would like to do. Currently client-capabilities are not a part of the `optimade-python-tools` repository. However, as this is not the first time this question has been raised for us, we have also started discussing how to deal with this internally - up to this point we have decided to dedicate other repositories for this endeavour and have the `optimade-python-tools` focus on database providers; how they can implement the OPTIMADE API specification as easily as possible and then how to handle the data items that are returned (as described above using the `optimade.adapters` module).

This is a rather long answer, and I am sorry for that. I think it clearly shows we need to document this motivational part of the repository better. And also perhaps the intended and actual use cases for the `optimade.adapters` module.
Furthermore, as OPTIMADE database providers finish implementing their servers, the need for client tools become more urgent. This is an extremely important point for us for the future of the repository, so thank you for highlighting this.

Again, thank you for your review and time Hung (@hungpham2017).
