### 2021-09-13 #3458 @ml-evs
Great news, assuming this is okay with @jgostick!!

Thanks again @jamesrhester and @hungpham2017 for volunteering your time to review our repository; we will be sure to take your feedback on board going forward. The tutorial-style guide that we wrote as a result of this review has already proved very useful and it has enabled at least two new OPTIMADE database implementations using our package as a result.

I have just made minor modifications to our affiliations, added a joint first authorship note with @CasperWA and added myself as corresponding author to the paper (as footnotes, following the example paper on the JOSS website).

If this is finalized, we will merge the PR https://github.com/Materials-Consortia/optimade-python-tools/pull/804 into our main branch and make a new release and archive v0.16.4 to be used as the final version for this paper.
