### 2021-09-15 #3458 @ml-evs
I have removed my email address footnote and replaced it with "corresponding author", as I think this was causing the PDF build failure. Interestingly, the PDF build did work with the whedon GH action (see https://github.com/Materials-Consortia/optimade-python-tools/runs/3600587073?check_suite_focus=true - can raise an issue if desired).

Hopefully everything is okay now.
