### 2021-09-22 #3458 @ml-evs
> @ml-evs Please update the metadata in your Zenodo archive so the author list and title exactly match your JOSS article.

Hi @kthyng, is this a hard requirement? Our automatic Zenodo release archives predate this submission by a couple of years and include all historic contributors to the package.

We could release an individual archive for this version+paper, if necessary.
