### 2021-09-22 #3458 @kthyng
@ml-evs No it isn't a hard requirement — it gives a clear association of which software archive matches the JOSS publication to have a nice final package to point to.  The individual archive sounds like a nice way to go, but you can choose your preference.
