### 2021-09-22 #3458 @ml-evs
> @ml-evs No it isn't a hard requirement — it gives a clear association of which software archive matches the JOSS publication to have a nice final package to point to. The individual archive sounds like a nice way to go, but you can choose your preference.

In that case, I think we would prefer to keep the single point of truth as the current archive, without changing the metadata (which would just mean removing people) - this means that the links to later versions are kept - assuming that is okay with you.
