### 2021-09-22 df5a149
Use single precision floating point machine epsilon for element_ratios
