### 2021-09-22 902157e
Clarify error messages for optional queries when server responds with 501
